On branch pa2  Changes to be committed: 	modified:   nemu/Kconfig 	modified:   npc/src/main/scala/One_cycle/EXU.scala 	modified:   npc/src/main/scala/One_cycle/IDU.scala 	modified:   npc/verilator/src/difftest/dut.c 	modified:   npc/verilator/src/include/difftest.h 	modified:   npc/verilator/src/include/tb.h 	modified:   npc/verilator/src/memory/pmem.c 	modified:   npc/verilator/src/tb_CPUTop.cpp 	modified:   npc/verilator/vsrc/CPUTop.v 	modified:   npc/verilator/vsrc/MEM.v

+++ npc/verilator/src/difftest/dut.c
@@ -14,7 +14,7 @@ Ref_difftest_exec ref_difftest_exec=NULL;
 Ref_difftest_raise_intr ref_difftest_raise_intr=NULL;
 Ref_difftest_init ref_difftest_init=NULL;
 
-void init_difftest(char *ref_so_file, uint32_t img_size, int port, uint32_t *mem){
+void init_difftest(char *ref_so_file, uint32_t img_size, int port, uint8_t *mem){
   assert(ref_so_file != NULL);
   
   void *handle;
